move filesystemapi to separate module
